WIP: improving span data when using preprocessor features
Since #if, macro expansion and friends all modify the source code we will loose track of which line the parser is looking at. So to fix this the preprocessor will generate source groups that have stable lines in it and we will later tokenize each group in order. Currently the preceeding groups don't get the correctt offset.